Create even_parity
For the square-well problem treated here and the choice of psi (0) for the even solutions, a cutoff parameter b is 1.5. The spatial step size was Delta_x=0.01. We have taken V=0 inside the well and V=V_o=1000 outside. The red line shows the position of the walls of the well.